restore detailed tag version format
This change reverts #186 and restores the use of detailed (dirty)
release version formatting.

contolini, Scotchester and I debugged this on the cfpb/retirement repo
and determined that the unexpected version numbers are due to the
frontend script changing the package-lock.json file.

See https://github.com/cfpb/retirement/pull/227 for the PR that fixed
this on the retirement repo.