Remove fixture loading from migrations
This commit removes all fixture data loading and manipulating done via Django
migrations. Project documentation is instead modified to encourage manual
loading of fixture data. Existing fixture data has been consolidated into a
single file.

This change fixes #327  which prevents the existing migrations from being run
on an empty Django database. It also brings this app in line with how the rest
of cf.gov data is loaded; that is, via some manual process as opposed to as
part of Django migrations.

Here are the existing migrations that alter database data, and how this change
impacts that data:

1. 0010_load_initial_data: deletes and loads data from
fixtures/tdp_initial_data.json for all TDP non-page models except for
ActivityBuildingBlock. This commit empties out this migration so that it
becomes a no-op.

2. 0011_update_initial_data: reloads data for the ActivityCouncilForEconEd
model from fixtures/tdp_updated_council_for_ed_data.json. Currently, the
content for this model in these two fixture files is identical, so this
migration appears to be redundant. This commit empties out this migration so
that it becomes a no-op.

3. 0014_update_building_blocks: loads data for the ActivityBuildingBlock model
from fixtures/tdp_updated_building_blocks.json. This commit empties out this
migration so that it becomes a no-op, and adds the ActivityBuildingBlock data
into tdp_initial_data.json.

4. 0016_auto_20180829_2323: updates database sequence IDs for the non-page
Django models. This migration only seems to have been necessary due to the
previous migrations; this commit empties it out so that it becomes a no-op.

5. 0018_update_student_characteristics: updates titles and weights of the
ActivityStudentCharacteristics model.

The original values in the tdp_initial_data.json file were:

  - pk: 1, weight: 0, title: "English language learner (ELL)"
  - pk: 2, weight: 0, title: "Free and reduced lunch (FRL)"
  - pk: 3, weight: 0, title: "Rural communities"
  - pk: 4, weight: 0, title: "Special education (SPED)"
  - pk: 5, weight: 0, title: "Urban communities"

The new values as set in migration `0018` are:

  - pk: 1, weight: 0, title: "English language learners"
  - pk: 2, weight: 2, title: "Low income"
  - pk: 3, weight: 3, title: "Rural"
  - pk: 4, weight: 1, title: "Special education"
  - pk: 5, weight: 4, title: "Urban"

This commit merges these changes back into the original tdp_initial_data.json
file, so the single fixture file reflects the updated values. It also empties
out the 0018 migration so that it becomes a no-op.

With all of these changes, the migrations for this project now follow the more
standard pattern of simply setting up the database tables, without trying to
to fill them with data.

Data can be loaded by using the standard loaddata command, for example in the
cfgov-refresh project:

$ cfgov/manage.py loaddata path/to/tdp/fixtures/tdp_initial_data.json

This commit updates the project INSTALL.md to document this recommendation.

This commit also modifies this project's test settings to ensure that
migrations are executed during test execution.