Don't try to install TDP into itself
cfgov-refresh PR 5083 aims to make the cf.gov satellite apps (including
this one) non-optional, meaning that they are treated the same as any
other Python dependency. This is a bit problematic for TDP, because TDP
both depends on and is depended on by cfgov-refresh.

This change modifies the install_cfgov_refresh.sh script so that it
filters out teachers_digital_platform when installing cfgov-refresh
dependencies. This avoids us installing a built wheel of this project
which would hide the local directories that we really want to test.