Use "current" and "future" for tox environments
This change replaces our version-specific factors for tox environments with a simpler `current` and `future` nomenclature.

This means to run the unittests on the current production version of Python, Django, and Wagtail, one would use:

```
tox -e unittest-current
```

To run them against upcoming versions of Python, Django, and Wagtail:

```
tox -e unittest-future
```

I've also added a convenience `unittest` environment which runs identically to `unittest-current`.

The intent here is to simplify the name of our test environments and make the naming version-independent.