Simplify the tox.ini
The tox.ini file in this repo used to mirror the cfgov-refresh tox.ini because TDP required cfgov-refresh to test, and testing was easier if the environments mirrored cfgov-refresh's. With the move of TDP to cfgov-refresh, that's no longer necessary and we can use a more traditional tox.ini file.